requirements.txt
Flask==3.0.3
PyMuPDF==1.24.10
python-dotenv==1.0.1
